+++ components/ImageList.tsx
@@ -1,10 +1,9 @@
-import { Card, SimpleGrid } from "@mantine/core";
-
+import { Button, Card } from "@mantine/core";
+import axios from "axios";
 import Link from "next/link";
-import { useState } from "react";
-import { useQuery } from "react-query";
-import { SdImage } from "../libs/shared-types/src";
+import { useQuery, useQueryClient } from "react-query";
 
+import { SdImage } from "../libs/shared-types/src";
 import { ImageTransformBuilder } from "./ImageTransform";
 import { SdImageComp } from "./SdImageComp";
 
@@ -13,6 +12,8 @@ export function getImageUrl(imageUrl: string): string {
 }
 
 export function ImageList() {
+  const qc = useQueryClient();
+
   const { data, isLoading, isError, error } = useQuery("images", async () => {
     const res = await fetch("/api/images");
     const results = (await res.json()) as SdImage[];
@@ -35,28 +36,44 @@ export function ImageList() {
     {}
   );
 
+  // function post a delete based on group id
+
+  const handleDelete = async (groupId: string) => {
+    // use axios for post
+    console.log("delete", groupId);
+    const res = await axios.delete(`/api/group/${groupId}`);
+
+    // invalidate the query
+    qc.invalidateQueries("images");
+  };
+
   return (
     <div>
       <h1>image groups</h1>
       <div>
         {isLoading ? "loading..." : ""}
         {isError ? "error" : ""}
-        <SimpleGrid cols={4}>
+        <div style={{ display: "flex", flexWrap: "wrap" }}>
           {(Object.keys(imageGroups) ?? []).map((id) => {
             const group = imageGroups[id];
             const img = group[0];
             return (
-              <Card key={img.id}>
+              <Card key={img.id} style={{ border: "1px solid black" }}>
                 <Link href={`/group/${img.groupId}`}>
                   <div>
                     <SdImageComp image={img} size={200} disablePopover />
                     <p>total items = {group.length}</p>
                   </div>
                 </Link>
+                <div>
+                  <Button onClick={() => handleDelete(img.groupId)}>
+                    delete...
+                  </Button>
+                </div>
               </Card>
             );
           })}
-        </SimpleGrid>
+        </div>
 
         <ImageTransformBuilder />
       </div>